requirements.txt
jinja2
playwright
pymysql
pyyaml
requests
